refactor: restructure component file organisation
- Moved components into dedicated subdirectories for improved modularity.
- Updated imports throughout the app to reflect new file paths.
- Updated test file locations to match the new component structure.
- Added `data-testid` attributes to theme toggle icons for better testability.
