+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Hero from "@/components/Hero/Hero";
-import About from "@/components/About";
+import About from "@/components/About/About";
 import Projects from "@/components/Projects/Projects";
-import Skills from "@/components/Skills";
+import Skills from "@/components/Skills/Skills";
 import Contact from "@/components/Contact/Contact";
 
 export default function Home() {
