+++ src/components/Footer/Footer.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import Footer from "@/components/Footer";
+import Footer from "@/components/Footer/Footer";
 
 describe("Footer component", () => {
   test("renders main text", () => {
