+++ src/components/ScrollToTopButton/ScrollToTopButton.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { act, fireEvent, render } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import ScrollToTopButton from '../components/ScrollToTopButton'
+import ScrollToTopButton from './ScrollToTopButton'
 
 describe('ScrollToTopButton', () => {
   beforeEach(() => {
