+++ src/components/ToggleTheme/ToggleTheme.tsx
@@ -5,7 +5,7 @@ import { Moon, Sun } from 'lucide-react'
 import { useTheme } from 'next-themes'
 
 const ToggleTheme: React.FC = () => {
-  const { setTheme, resolvedTheme } = useTheme()
+  const {setTheme, resolvedTheme} = useTheme()
   const [mounted, setMounted] = useState(false)
   
   useEffect(() => {
@@ -25,8 +25,8 @@ const ToggleTheme: React.FC = () => {
       aria-label="Toggle theme"
     >
       {resolvedTheme === 'dark'
-        ? <Sun className="w-5 h-5" />
-        : <Moon className="w-5 h-5" />
+        ? <Sun className="w-5 h-5" data-testid="sun-icon"/>
+        : <Moon className="w-5 h-5" data-testid="moon-icon"/>
       }
     </button>
   )
